Fix hellspider generation crashing.
In run_hs, int h was not initialized, which made it random. I'm assuming the compiler used previously was being friendly and set it to zero. There's always just one hellspider, so h is not actually used in a meaningful way. I assume this is legacy code from a previous game with more than one enemy at a time.